pyproject.toml
[tool.poetry]
name = "intelligent-agent-project"
version = "0.1.0"
description = "An intelligent AI agent framework built with Python"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
homepage = ""
repository = ""
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
python-dotenv = ">=1.0.0"
pydantic = ">=2.0.0"
pydantic-settings = ">=2.0.0"
pydantic-core = ">=2.0.0"
annotated-types = ">=0.7.0"
openai = ">=1.0.0"
anthropic = ">=0.5.0"
fastapi = ">=0.100.0"
uvicorn = ">=0.23.0"
sqlalchemy = ">=2.0.0"
httpx = ">=0.27.0"
requests = ">=2.31.0"
requests-toolbelt = ">=1.0.0"
starlette = ">=0.50.0"
anyio = ">=4.0.0"
aiohttp = ">=3.9.0"
orjson = ">=3.10.0"
tiktoken = ">=0.12.0"
tenacity = ">=9.0.0"
PyYAML = ">=6.0.0"
typing-extensions = ">=4.0.0"
urllib3 = ">=2.0.0"
certifi = ">=2024.0.0"
charset-normalizer = ">=3.0.0"
idna = ">=3.0"
jiter = ">=0.12.0"
multidict = ">=6.0.0"
frozenlist = ">=1.4.0"
aiosignal = ">=1.3.0"
aiohappyeyeballs = ">=2.3.0"
attrs = ">=23.0.0"
propcache = ">=0.4.0"
yarl = ">=1.9.0"
sniffio = ">=1.3.0"
h11 = ">=0.16.0"
httpcore = ">=1.0.0"
greenlet = ">=3.0.0"
regex = ">=2023.0.0"
pytokens = ">=0.3.0"
jsonpatch = ">=1.33.0"
jsonpointer = ">=3.0.0"
colorama = ">=0.4.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.0.0"
pytest-cov = ">=4.0.0"
pre-commit = ">=3.0.0"
black = ">=23.0.0"
isort = ">=5.12.0"
flake8 = ">=6.0.0"
mypy = ">=1.0.0"
mkdocs = ">=1.5.0"
mkdocs-material = ">=9.0.0"
click = ">=8.0.0"
filelock = ">=3.0.0"
platformdirs = ">=4.0.0"
virtualenv = ">=20.0.0"
cfgv = ">=3.0.0"
distlib = ">=0.3.0"
identify = ">=2.0.0"
nodeenv = ">=1.0.0"
packaging = ">=23.0"
pathspec = ">=0.12.0"
pycodestyle = ">=2.14.0"
pyflakes = ">=3.4.0"
mccabe = ">=0.7.0"
setuptools = ">=68.0.0"
wheel = ">=0.45.0"
tqdm = ">=4.0.0"
typing-inspection = ">=0.4.0"
xxhash = ">=3.0.0"

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(\.
  |build
  |dist
  |venv
  |\.venv
  |tests
  |examples
  |docs
  |scripts
  |data
  |notebooks
)/
'''

[tool.isort]
profile = "black"
line_length = 88
extend_skip_glob = ["tests/**", "examples/**", "docs/**", "scripts/**", "data/**", "notebooks/**"]

[tool.mypy]
python_version = "3.10"
warn_unused_configs = true
ignore_missing_imports = true
follow_imports = "silent"
exclude = [
  '^src/desktop-tauri/node_modules/',
  '^src/desktop-tauri/',
  '^data/',
  '^scripts/',
]
strict = false
disable_error_code = ["override"]

[tool.pytest.ini_options]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
testpaths = ["tests"]
addopts = ["-v", "--cov=src", "--cov-report=term-missing"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
